Header only, transforming some functions to inline, created functions.cpp, created CMakeLists.txt
